start organizing unit tests; fix package name
