+++ pkg/core/actions/getnextimage_test.go
@@ -22,16 +22,23 @@ under the License.
 package actions
 
 import (
+	"testing"
+
 	m "github.com/blackducksoftware/perceptor/pkg/core/model"
+	log "github.com/sirupsen/logrus"
 )
 
-type AllPods struct {
-	Pods []m.Pod
-}
+func TestGetNextImageForScanningActionNoImageAvailable(t *testing.T) {
+	// actual
+	var nextImage *m.Image
+	actual := m.NewModel(&m.Config{ConcurrentScanLimit: 3}, "test version")
+	(&GetNextImage{func(image *m.Image) {
+		nextImage = image
+	}}).Apply(actual)
+	// expected: front image removed from scan queue, status and time of image changed
+	expected := *m.NewModel(&m.Config{ConcurrentScanLimit: 3}, "test version")
 
-func (a *AllPods) Apply(model *m.Model) {
-	model.Pods = map[string]m.Pod{}
-	for _, pod := range a.Pods {
-		model.AddPod(pod)
-	}
+	assertEqual(t, nextImage, nil)
+	log.Infof("%+v, %+v", actual, expected)
+	// assertEqual(t, actual, expected)
 }
