+++ pkg/core/actions/getscanresults.go
@@ -19,7 +19,7 @@ specific language governing permissions and limitations
 under the License.
 */
 
-package core
+package actions
 
 import (
 	"github.com/blackducksoftware/perceptor/pkg/api"
